requirements.txt
tensorflow-macos==2.13.0
keras==2.13.1
keras-cv==0.6.1
matplotlib==3.7.2
numpy==1.24.3
pandas==2.0.3
optuna==3.5.0
pillow==10.0.0
tqdm==4.65.0